Fixed English translations
